Port NamedTempFile, TempFile, and platform file operations

+++ src/linuxMain/kotlin/io/github/kotlinmania/tempfile/dir/MkdirLinux.kt
@@ -85,6 +85,8 @@ private fun posixLstatIsDir(path: String): Boolean? =
         val sb = alloc<stat>()
         val rc = stat(path, sb.ptr)
         if (rc != 0) return@memScoped null
-        val mode = sb.st_mode
-        (mode and S_IFDIR) != 0u
+        val mode = sb.st_mode.toInt()
+        (mode and S_IFDIR) != 0
     }
+
+
